DiscretizedScale: Move class definition from scatter plot to plotutils
